+++ rhythm-one-htb-validator.js
@@ -33,22 +33,6 @@ function partnerValidator(configs) {
                                 type: 'string',
                                 minLength: 1
                             },
-                            zone: {
-                                type: 'string',
-                                minLength: 1
-                            },
-                            path: {
-                                type: 'string',
-                                minLength: 1
-                            },
-                            imp: {
-                                type: 'string',
-                                minLength: 1
-                            },
-                            adType: {
-                                type: 'string',
-                                minLength: 1
-                            },
                             sizes: {
                                 type: 'array',
                                 items: {