+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 2.0.2
+- Updated query string parameter name for GDPR
+- Removed optional parameters from rhythm-one-htb-validator.js file
+
 # 2.0.1
 - Updated review comments, changed window reference to Browser.topWindow